+++ src/textual/dom.py
@@ -208,6 +208,11 @@ class DOMNode(MessagePump):
             )
         self._auto_refresh = interval
 
+    @property
+    def is_modal(self) -> bool:
+        """Is the node a modal?"""
+        return False
+
     def _automatic_refresh(self) -> None:
         """Perform an automatic refresh (set with auto_refresh property)."""
         self.refresh()