+++ CHANGELOG.md
@@ -5,12 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.17.2] - 2023-04-02
 
-### Fixed
+### [Fixed]
 
 - Fixed bindings persistance https://github.com/Textualize/textual/issues/1613
 - The `Markdown` widget now auto-increments ordered lists https://github.com/Textualize/textual/issues/2002
+- Fixed modal bindings https://github.com/Textualize/textual/issues/2194
+- Fix binding enter to active button https://github.com/Textualize/textual/issues/2194
+
+### [Changed]
+
+- tab and shift+tab are now defined on Screen.
 
 ## [0.17.1] - 2023-03-30
 
@@ -687,6 +693,9 @@ https://textual.textualize.io/blog/2022/11/08/version-040/#version-040
 - New handler system for messages that doesn't require inheritance
 - Improved traceback handling
 
+[0.17.2]: https://github.com/Textualize/textual/compare/v0.17.1...v0.17.2
+[0.17.1]: https://github.com/Textualize/textual/compare/v0.17.0...v0.17.1
+[0.17.0]: https://github.com/Textualize/textual/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/Textualize/textual/compare/v0.15.1...v0.16.0
 [0.15.1]: https://github.com/Textualize/textual/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/Textualize/textual/compare/v0.14.0...v0.15.0